move utils/2bttns.tsx out of pages
